backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
firecrawl-py==0.0.16
fpdf2==2.7.6
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0

